Implemented Photonk
Implemented the Photonk (Photo-link...ish) library. Basically a clickable
photo that copies text to your clipboard. It's still rough but it's mostly
all there - Right now it currently makes you 'manually' copy the url
(which isn't a full url either) - but I plan on switching to the textarea
method soon and firing a klaxon to notify the link was copied.